Make sure MBWMObjectClass::type is big enough to hold GType
On 64 bit platforms size of GType is 64 bits, so int is not big enough to
hold GType.

+++ matchbox/core/mb-wm-object.h
@@ -66,7 +66,7 @@ MBWMObjectClassInfo;
  */
 struct MBWMObjectClass
 {
-  int              type;
+  unsigned long    type;
   MBWMObjectClass *parent;
   MBWMObjVargFunc  init;
   MBWMObjFunc      destroy;
